Version 4.2 Implemented a NSTimer to decrease the time left variable every second. Store the time set by the difficulty choosen.

+++ Whack-a-mole/ScoreDataClass.h
@@ -13,6 +13,6 @@
 #pragma mark properties
 
 @property int currentscore;
-@property int highestscore;
+@property NSInteger highestscore;
 
 @end
